Refactored ErrorResponse, moved from error package to response package
