+++ src/main/java/org/java2uml/java2umlapi/restControllers/LWControllers/BodyController.java
@@ -11,8 +11,8 @@ import org.java2uml.java2umlapi.lightWeight.Body;
 import org.java2uml.java2umlapi.lightWeight.repository.BodyRepository;
 import org.java2uml.java2umlapi.lightWeight.repository.LightWeightRepository;
 import org.java2uml.java2umlapi.modelAssemblers.BodyAssembler;
-import org.java2uml.java2umlapi.restControllers.error.ErrorResponse;
 import org.java2uml.java2umlapi.restControllers.exceptions.LightWeightNotFoundException;
+import org.java2uml.java2umlapi.restControllers.response.ErrorResponse;
 import org.springframework.hateoas.EntityModel;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
