+++ src/main/java/org/java2uml/java2umlapi/restControllers/LWControllers/ClassOrInterfaceController.java
@@ -11,8 +11,8 @@ import org.java2uml.java2umlapi.lightWeight.ClassOrInterface;
 import org.java2uml.java2umlapi.lightWeight.repository.ClassOrInterfaceRepository;
 import org.java2uml.java2umlapi.lightWeight.repository.SourceRepository;
 import org.java2uml.java2umlapi.modelAssemblers.ClassOrInterfaceAssembler;
-import org.java2uml.java2umlapi.restControllers.error.ErrorResponse;
 import org.java2uml.java2umlapi.restControllers.exceptions.LightWeightNotFoundException;
+import org.java2uml.java2umlapi.restControllers.response.ErrorResponse;
 import org.springframework.hateoas.CollectionModel;
 import org.springframework.hateoas.EntityModel;
 import org.springframework.web.bind.annotation.GetMapping;
