+++ src/main/java/org/java2uml/java2umlapi/restControllers/LWControllers/ClassRelationController.java
@@ -11,9 +11,9 @@ import org.java2uml.java2umlapi.lightWeight.ClassRelation;
 import org.java2uml.java2umlapi.lightWeight.repository.ClassRelationRepository;
 import org.java2uml.java2umlapi.lightWeight.repository.SourceRepository;
 import org.java2uml.java2umlapi.modelAssemblers.ClassRelationAssembler;
-import org.java2uml.java2umlapi.restControllers.error.ErrorResponse;
 import org.java2uml.java2umlapi.restControllers.exceptions.ClassRelationNotFoundException;
 import org.java2uml.java2umlapi.restControllers.exceptions.LightWeightNotFoundException;
+import org.java2uml.java2umlapi.restControllers.response.ErrorResponse;
 import org.springframework.hateoas.CollectionModel;
 import org.springframework.hateoas.EntityModel;
 import org.springframework.web.bind.annotation.GetMapping;
