+++ src/main/java/org/java2uml/java2umlapi/restControllers/LWControllers/ConstructorController.java
@@ -13,8 +13,8 @@ import org.java2uml.java2umlapi.lightWeight.repository.ConstructorRepository;
 import org.java2uml.java2umlapi.lightWeight.repository.LightWeightRepository;
 import org.java2uml.java2umlapi.modelAssemblers.ConstructorAssembler;
 import org.java2uml.java2umlapi.restControllers.SwaggerDescription;
-import org.java2uml.java2umlapi.restControllers.error.ErrorResponse;
 import org.java2uml.java2umlapi.restControllers.exceptions.LightWeightNotFoundException;
+import org.java2uml.java2umlapi.restControllers.response.ErrorResponse;
 import org.springframework.hateoas.CollectionModel;
 import org.springframework.hateoas.EntityModel;
 import org.springframework.web.bind.annotation.GetMapping;
