+++ src/main/java/org/java2uml/java2umlapi/restControllers/LWControllers/EnumConstantController.java
@@ -11,8 +11,8 @@ import org.java2uml.java2umlapi.lightWeight.EnumConstant;
 import org.java2uml.java2umlapi.lightWeight.repository.EnumConstantRepository;
 import org.java2uml.java2umlapi.lightWeight.repository.EnumLWRepository;
 import org.java2uml.java2umlapi.modelAssemblers.EnumConstantAssembler;
-import org.java2uml.java2umlapi.restControllers.error.ErrorResponse;
 import org.java2uml.java2umlapi.restControllers.exceptions.LightWeightNotFoundException;
+import org.java2uml.java2umlapi.restControllers.response.ErrorResponse;
 import org.springframework.hateoas.CollectionModel;
 import org.springframework.hateoas.EntityModel;
 import org.springframework.web.bind.annotation.GetMapping;
