+++ src/main/java/org/java2uml/java2umlapi/restControllers/LWControllers/EnumLWController.java
@@ -11,8 +11,8 @@ import org.java2uml.java2umlapi.lightWeight.EnumLW;
 import org.java2uml.java2umlapi.lightWeight.repository.EnumLWRepository;
 import org.java2uml.java2umlapi.lightWeight.repository.SourceRepository;
 import org.java2uml.java2umlapi.modelAssemblers.EnumLWAssembler;
-import org.java2uml.java2umlapi.restControllers.error.ErrorResponse;
 import org.java2uml.java2umlapi.restControllers.exceptions.LightWeightNotFoundException;
+import org.java2uml.java2umlapi.restControllers.response.ErrorResponse;
 import org.springframework.hateoas.CollectionModel;
 import org.springframework.hateoas.EntityModel;
 import org.springframework.web.bind.annotation.GetMapping;
