+++ src/main/java/org/java2uml/java2umlapi/restControllers/LWControllers/MethodController.java
@@ -12,8 +12,8 @@ import org.java2uml.java2umlapi.lightWeight.Method;
 import org.java2uml.java2umlapi.lightWeight.repository.LightWeightRepository;
 import org.java2uml.java2umlapi.lightWeight.repository.MethodRepository;
 import org.java2uml.java2umlapi.modelAssemblers.MethodAssembler;
-import org.java2uml.java2umlapi.restControllers.error.ErrorResponse;
 import org.java2uml.java2umlapi.restControllers.exceptions.LightWeightNotFoundException;
+import org.java2uml.java2umlapi.restControllers.response.ErrorResponse;
 import org.springframework.hateoas.CollectionModel;
 import org.springframework.hateoas.EntityModel;
 import org.springframework.web.bind.annotation.GetMapping;
